+++ frontend/.eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\App.tsx":"1"},{"size":1661,"mtime":1608916754848,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo","C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\App.tsx",[]]
+[{"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx":"1","C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts":"2","C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx":"3"},{"size":506,"mtime":1608916771827,"results":"4","hashOfConfig":"5"},{"size":425,"mtime":1608916282147,"results":"6","hashOfConfig":"5"},{"size":2132,"mtime":1609231944118,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"xlb0qo",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\karan\\git\\book-project\\frontend\\src\\index.tsx",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\reportWebVitals.ts",[],"C:\\Users\\karan\\git\\book-project\\frontend\\src\\login\\Login.tsx",[]]
