+++ frontend/public/index.html
@@ -27,6 +27,8 @@
     <title>Book Project</title>
     <!-- Font for Material UI -->
     <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
+    <!-- Material icons -->
+    <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" />
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
