+++ frontend/src/App.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Login from "./login/Login";
+
+function App() {
+    return (
+        <Login />
+    )
+}
+
+export default App;
